refactor(ivf): split implementation by responsibility
