+++ src/algorithm/ivf/ivf.cpp
@@ -1,4 +1,3 @@
-
 // Copyright 2024-present the vsag project
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,6 +14,8 @@
 
 #include "ivf.h"
 
+#include <fmt/format.h>
+
 #include <algorithm>
 #include <atomic>
 #include <cmath>
@@ -26,43 +27,22 @@
 #include <unordered_map>
 
 #include "algorithm/inner_index_interface.h"
-#include "attr/argparse.h"
-#include "attr/executor/executor.h"
 #include "datacell/bucket_datacell_parameter.h"
 #include "datacell/flatten_datacell_parameter.h"
 #include "datacell/flatten_interface.h"
-#include "datacell/graph_datacell_parameter.h"
-#include "datacell/graph_interface_parameter.h"
 #include "flat_bucket_searcher.h"
 #include "gno_imi_partition.h"
 #include "graph_bucket_searcher.h"
-#include "impl/heap/standard_heap.h"
-#include "impl/inner_search_param.h"
-#include "impl/pruning_strategy.h"
-#include "impl/reasoning/search_reasoning.h"
 #include "impl/reorder/bucket_reorder.h"
 #include "impl/reorder/flatten_reorder.h"
-#include "impl/searcher/basic_searcher.h"
 #include "index/index_impl.h"
-#include "index_feature_list.h"
 #include "inner_string_params.h"
-#include "io/reader_io/reader_io_parameter.h"
 #include "ivf_nearest_partition.h"
-#include "query_context.h"
-#include "simd/normalize.h"
-#include "storage/serialization.h"
-#include "storage/serialization_tags.h"
-#include "storage/stream_reader.h"
 #include "storage/stream_writer.h"
-#include "storage/tlv_section.h"
-#include "utils/search_threshold.h"
 #include "utils/util_functions.h"
-#include "utils/visited_list.h"
 #include "vsag_exception.h"
 
 namespace vsag {
-
-static constexpr BucketIdType INVALID_BUCKET_ID = static_cast<BucketIdType>(-1);
 
 namespace {
 
@@ -305,451 +285,6 @@ IVF::GetCodeByInnerId(InnerIdType inner_id, uint8_t* data) const {
     this->bucket_->GetCodesById(bucket_id, offset_id, data);
 }
 
-void
-IVF::InitFeatures() {
-    // Common Init
-    // Build & Add
-    this->index_feature_list_->SetFeatures({
-        IndexFeature::SUPPORT_BUILD,
-        IndexFeature::SUPPORT_ADD_AFTER_BUILD,
-        IndexFeature::SUPPORT_ADD_CONCURRENT,
-    });
-
-    // search
-    this->index_feature_list_->SetFeatures({
-        IndexFeature::SUPPORT_KNN_SEARCH,
-        IndexFeature::SUPPORT_KNN_SEARCH_WITH_ID_FILTER,
-    });
-    // concurrency
-    this->index_feature_list_->SetFeature(IndexFeature::SUPPORT_SEARCH_CONCURRENT);
-
-    // serialize
-    this->index_feature_list_->SetFeatures({
-        IndexFeature::SUPPORT_DESERIALIZE_BINARY_SET,
-        IndexFeature::SUPPORT_DESERIALIZE_FILE,
-        IndexFeature::SUPPORT_DESERIALIZE_READER_SET,
-        IndexFeature::SUPPORT_SERIALIZE_BINARY_SET,
-        IndexFeature::SUPPORT_SERIALIZE_FILE,
-        IndexFeature::SUPPORT_SERIALIZE_WRITE_FUNC,
-    });
-
-    auto name = this->bucket_->GetQuantizerName();
-    if (name != QUANTIZATION_TYPE_VALUE_FP32 and name != QUANTIZATION_TYPE_VALUE_BF16 and
-        name != QUANTIZATION_TYPE_VALUE_FP16) {
-        this->index_feature_list_->SetFeature(IndexFeature::NEED_TRAIN);
-    } else {
-        this->index_feature_list_->SetFeatures({
-            IndexFeature::SUPPORT_RANGE_SEARCH,
-            IndexFeature::SUPPORT_RANGE_SEARCH_WITH_ID_FILTER,
-        });
-    }
-
-    bool has_fp32 = false;
-    if (use_reorder_) {
-        const auto precise_quantizer_name = precise_bucket_ != nullptr
-                                                ? precise_bucket_->GetQuantizerName()
-                                                : reorder_codes_->GetQuantizerName();
-        has_fp32 = precise_quantizer_name == QUANTIZATION_TYPE_VALUE_FP32;
-    }
-    if (name == QUANTIZATION_TYPE_VALUE_FP32 or has_fp32) {
-        this->index_feature_list_->SetFeature(IndexFeature::SUPPORT_CAL_DISTANCE_BY_ID);
-        this->index_feature_list_->SetFeature(IndexFeature::SUPPORT_BATCH_CALC_DISTANCE_BY_ID);
-    }
-
-    if (name == QUANTIZATION_TYPE_VALUE_FP32 and
-        this->bucket_->GetMetricType() != MetricType::METRIC_TYPE_COSINE and
-        not bucket_->UseResidual()) {
-        this->index_feature_list_->SetFeature(IndexFeature::SUPPORT_GET_DATA_BY_IDS);
-    }
-
-    this->index_feature_list_->SetFeatures({IndexFeature::SUPPORT_CLONE,
-                                            IndexFeature::SUPPORT_EXPORT_MODEL,
-                                            IndexFeature::SUPPORT_GET_MEMORY_USAGE,
-                                            IndexFeature::SUPPORT_MERGE_INDEX});
-    if (this->bucket_->GetQuantizerName() == QUANTIZATION_TYPE_VALUE_PQFS) {
-        this->index_feature_list_->SetFeature(IndexFeature::SUPPORT_ADD_AFTER_BUILD, false);
-    }
-}
-
-std::vector<int64_t>
-IVF::Build(const DatasetPtr& base) {
-    if (graph_build_threshold_ > 0) {
-        CHECK_ARGUMENT(this->total_elements_ == 0,
-                       "graph bucket searcher requires a fresh Build with no prior data");
-    }
-    this->Train(base);
-    // TODO(LHT): duplicate
-    auto result = this->Add(base);
-    if (graph_build_threshold_ > 0) {
-        this->build_bucket_graphs();
-    }
-    this->cal_memory_usage();
-    return result;
-}
-
-void
-IVF::Train(const DatasetPtr& data) {
-    if (this->is_trained_) {
-        return;
-    }
-
-    int64_t total_elements = data->GetNumElements();
-    int64_t dim = data->GetDim();
-    DatasetPtr train_data =
-        vsag::sample_train_data(data, total_elements, dim, train_sample_count_, allocator_);
-    int64_t sample_count = train_data->GetNumElements();
-
-    partition_strategy_->Train(train_data);
-
-    const auto* data_ptr = train_data->GetFloat32Vectors();
-    this->bucket_->Train(data_ptr, sample_count);
-    if (use_reorder_) {
-        if (precise_bucket_ != nullptr) {
-            this->precise_bucket_->Train(data->GetFloat32Vectors(), data->GetNumElements());
-        } else {
-            this->reorder_codes_->Train(data->GetFloat32Vectors(), data->GetNumElements());
-        }
-    }
-    this->is_trained_ = true;
-}
-
-std::vector<int64_t>
-IVF::Add(const DatasetPtr& base) {
-    // TODO(LHT): duplicate
-    if (not partition_strategy_->is_trained_) {
-        throw VsagException(ErrorType::INTERNAL_ERROR, "ivf index add without train error");
-    }
-    this->bucket_->Unpack();
-    if (precise_bucket_ != nullptr) {
-        this->precise_bucket_->Unpack();
-    }
-    auto num_element = base->GetNumElements();
-    const auto* ids = base->GetIds();
-    const auto* vectors = base->GetFloat32Vectors();
-    const auto* attr_sets = base->GetAttributeSets();
-    const auto* extra_info = base->GetExtraInfos();
-    const auto extra_info_size = base->GetExtraInfoSize();
-    auto buckets =
-        partition_strategy_->ClassifyDatas(vectors, num_element, buckets_per_data_, nullptr);
-
-    int64_t current_num;
-    bool need_cal_memory_usage = false;
-    {
-        std::lock_guard lock(label_lookup_mutex_);
-        current_num = this->total_elements_;
-        if (precise_bucket_ != nullptr) {
-            if (num_element < 0 or current_num < 0) {
-                throw VsagException(ErrorType::INVALID_ARGUMENT,
-                                    "invalid IVF precise bucket element count");
-            }
-            const auto posting_count = static_cast<uint64_t>(num_element);
-            const auto current_count = static_cast<uint64_t>(current_num);
-            const auto max_inner_id =
-                static_cast<uint64_t>(std::numeric_limits<InnerIdType>::max());
-            if (posting_count > max_inner_id or current_count > max_inner_id - posting_count) {
-                throw VsagException(ErrorType::INVALID_ARGUMENT,
-                                    "IVF precise bucket batch exceeds inner id capacity");
-            }
-        }
-        if (use_reorder_ and precise_bucket_ == nullptr) {
-            this->reorder_codes_->BatchInsertVector(base->GetFloat32Vectors(),
-                                                    base->GetNumElements());
-        }
-        for (int64_t i = 0; i < num_element; ++i) {
-            this->label_table_->Insert(i + total_elements_, ids[i]);
-        }
-        this->total_elements_ += num_element;
-        if (this->total_elements_ - last_cal_memory_element_ >= cal_memory_element_interval_) {
-            need_cal_memory_usage = true;
-            last_cal_memory_element_ = this->total_elements_;
-        }
-        location_map_.resize(this->total_elements_);
-    }
-
-    Vector<InnerIdType> precise_offsets(allocator_);
-    if (precise_bucket_ != nullptr) {
-        const auto posting_count = static_cast<uint64_t>(num_element);
-        Vector<InnerIdType> posting_ids(posting_count, allocator_);
-        precise_offsets.resize(posting_count);
-        for (uint64_t i = 0; i < posting_count; ++i) {
-            posting_ids[i] = static_cast<InnerIdType>(i + static_cast<uint64_t>(current_num));
-        }
-        precise_bucket_->BatchInsertVector(vectors,
-                                           buckets.data(),
-                                           posting_ids.data(),
-                                           static_cast<InnerIdType>(posting_count),
-                                           precise_offsets.data());
-    }
-
-    auto add_func = [&](int64_t i) -> void {
-        for (int64_t j = 0; j < buckets_per_data_; ++j) {
-            const auto* data_ptr = vectors + i * dim_;
-            auto idx = i * buckets_per_data_ + j;
-            auto posting_id = static_cast<InnerIdType>(idx + current_num * buckets_per_data_);
-            InnerIdType offset_id;
-            if (precise_bucket_ != nullptr) {
-                // Publish the basic posting only after its precise mirror is ready.
-                offset_id = precise_offsets[idx];
-                bucket_->InsertVectorWithOffset(data_ptr, buckets[idx], posting_id, offset_id);
-            } else {
-                offset_id = bucket_->InsertVector(data_ptr, buckets[idx], posting_id);
-            }
-            if (j == 0) {
-                std::lock_guard lock(label_lookup_mutex_);
-                location_map_[i + current_num] =
-                    (static_cast<uint64_t>(buckets[idx]) << LOCATION_SPLIT_BIT) |
-                    static_cast<uint64_t>(offset_id);
-            }
-            if (use_attribute_filter_ and this->attr_filter_index_ != nullptr and
-                attr_sets != nullptr) {
-                const auto& attr_set = attr_sets[i];
-                this->attr_filter_index_->Insert(attr_set, offset_id, buckets[idx]);
-            }
-            if (extra_info_size > 0) {
-                this->extra_infos_->InsertExtraInfo(extra_info + i * extra_info_size,
-                                                    i + current_num);
-            }
-        }
-    };
-    std::vector<std::future<void>> futures;
-    std::exception_ptr first_exception = nullptr;
-    try {
-        for (int64_t i = 0; i < num_element; ++i) {
-            if (this->thread_pool_ != nullptr) {
-                futures.emplace_back(thread_pool_->GeneralEnqueue(add_func, i));
-            } else {
-                add_func(i);
-            }
-        }
-    } catch (...) {
-        first_exception = std::current_exception();
-    }
-
-    if (this->thread_pool_ != nullptr) {
-        for (auto& future : futures) {
-            try {
-                future.get();
-            } catch (...) {
-                if (first_exception == nullptr) {
-                    first_exception = std::current_exception();
-                }
-            }
-        }
-    }
-    if (first_exception != nullptr) {
-        std::rethrow_exception(first_exception);
-    }
-    this->bucket_->Package();
-    if (precise_bucket_ != nullptr) {
-        this->precise_bucket_->Package();
-    }
-    if (need_cal_memory_usage) {
-        this->cal_memory_usage();
-    }
-    return {};
-}
-
-static GraphInterfaceParamPtr
-get_ivf_graph_param(const std::string& graph_param_string) {
-    auto graph_json = JsonType::Parse(graph_param_string);
-    if (!graph_json.Contains(IO_PARAMS_KEY) || !graph_json.Contains(GRAPH_STORAGE_TYPE_KEY) ||
-        graph_json[GRAPH_STORAGE_TYPE_KEY].GetString() != GRAPH_STORAGE_TYPE_VALUE_FLAT) {
-        throw VsagException(ErrorType::INVALID_BINARY,
-                            "bucket graph requires flat storage with memory-backed IO");
-    }
-
-    auto param = std::dynamic_pointer_cast<GraphDataCellParameter>(
-        GraphInterfaceParameter::GetGraphParameterByJson(
-            GraphStorageTypes::GRAPH_STORAGE_TYPE_VALUE_FLAT, graph_json));
-    if (param == nullptr || param->io_parameter_ == nullptr || param->support_remove_) {
-        throw VsagException(ErrorType::INVALID_BINARY, "invalid bucket graph parameters");
-    }
-
-    const auto io_type = param->io_parameter_->GetTypeName();
-    if (io_type != IO_TYPE_VALUE_MEMORY_IO && io_type != IO_TYPE_VALUE_BLOCK_MEMORY_IO) {
-        throw VsagException(ErrorType::INVALID_BINARY,
-                            fmt::format("unsupported bucket graph IO type: {}", io_type));
-    }
-    return param;
-}
-
-static JsonType
-serialize_ivf_graph_param(const GraphInterfaceParamPtr& graph_param) {
-    auto graph_json = graph_param->ToJson();
-    graph_json[GRAPH_STORAGE_TYPE_KEY].SetString(GRAPH_STORAGE_TYPE_VALUE_FLAT);
-    return graph_json;
-}
-
-static bool
-has_any_fresh_bucket_graph(const BucketInterfacePtr& bucket,
-                           const Vector<GraphInterfacePtr>& bucket_graphs) {
-    for (BucketIdType b = 0; b < static_cast<BucketIdType>(bucket_graphs.size()); ++b) {
-        if (bucket_graphs[b] != nullptr &&
-            bucket_graphs[b]->TotalCount() == bucket->GetBucketSize(b)) {
-            return true;
-        }
-    }
-    return false;
-}
-
-class PairwiseBucketDistanceProvider final : public DistanceProviderForGraph {
-public:
-    PairwiseBucketDistanceProvider(std::shared_ptr<BucketInterface> bucket,
-                                   BucketIdType bucket_id,
-                                   InnerIdType query_id)
-        : bucket_(std::move(bucket)), bucket_id_(bucket_id), query_id_(query_id) {
-    }
-
-    [[nodiscard]] float
-    QueryDistance(InnerIdType id, QueryContext* ctx = nullptr) const override {
-        return bucket_->ComputePairVectors(bucket_id_, query_id_, id);
-    }
-
-    void
-    BatchQueryDistance(float* distances,
-                       const InnerIdType* ids,
-                       InnerIdType count,
-                       QueryContext* ctx = nullptr) const override {
-        for (InnerIdType i = 0; i < count; ++i) {
-            distances[i] = bucket_->ComputePairVectors(bucket_id_, query_id_, ids[i]);
-        }
-    }
-
-    [[nodiscard]] float
-    PairwiseDistance(InnerIdType id1,
-                     InnerIdType id2,
-                     const ComputerInterfacePtr& computer = nullptr) const override {
-        return bucket_->ComputePairVectors(bucket_id_, id1, id2);
-    }
-
-    [[nodiscard]] ComputerInterfacePtr
-    FactoryComputerById(InnerIdType id) const override {
-        throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                            "PairwiseBucketDistanceProvider does not support FactoryComputerById");
-    }
-
-private:
-    std::shared_ptr<BucketInterface> bucket_;
-    InnerIdType query_id_;
-    BucketIdType bucket_id_;
-};
-
-void
-IVF::build_bucket_graphs() {
-    if (graph_build_threshold_ <= 0) {
-        return;
-    }
-    if (graph_param_ == nullptr) {
-        graph_param_ = std::make_shared<GraphDataCellParameter>();
-    }
-
-    constexpr int64_t max_degree = 64;
-    constexpr uint64_t ef_construction = 300;
-    const auto bucket_count = bucket_->bucket_count_;
-    bucket_graphs_.resize(bucket_count);
-
-    auto build_one_bucket = [&](BucketIdType b) {
-        const auto bucket_size = bucket_->GetBucketSize(b);
-        if (bucket_size < graph_build_threshold_) {
-            return;
-        }
-
-        const auto* inner_ids = bucket_->GetInnerIds(b);
-        Vector<InnerIdType> valid_ids(allocator_);
-        valid_ids.reserve(bucket_size);
-        for (InnerIdType i = 0; i < static_cast<InnerIdType>(bucket_size); ++i) {
-            if (inner_ids[i] != std::numeric_limits<InnerIdType>::max()) {
-                valid_ids.push_back(i);
-            }
-        }
-        if (valid_ids.size() < static_cast<uint64_t>(graph_build_threshold_)) {
-            return;
-        }
-
-        const auto effective_degree =
-            std::min(max_degree, static_cast<int64_t>(valid_ids.size()) - 1);
-        if (effective_degree <= 0) {
-            return;
-        }
-
-        auto graph = GraphInterface::MakeInstance(graph_param_, this->common_param_);
-        graph->Resize(bucket_size);
-        graph->SetTotalCount(bucket_size);
-        graph->SetMaximumDegree(static_cast<uint32_t>(effective_degree));
-
-        auto mutexes = std::make_shared<EmptyMutex>();
-        BasicSearcher searcher(common_param_);
-
-        const auto entry = valid_ids.front();
-        graph->InsertNeighborsById(entry, Vector<InnerIdType>(allocator_));
-        auto visited = std::make_shared<VisitedList>(bucket_size, allocator_);
-        for (uint64_t node_pos = 1; node_pos < valid_ids.size(); ++node_pos) {
-            const auto node = valid_ids[node_pos];
-            InnerSearchParam search_param;
-            search_param.ep = entry;
-            search_param.ef = std::min(ef_construction, node_pos);
-            search_param.topk = static_cast<int64_t>(search_param.ef);
-            PairwiseBucketDistanceProvider distance_provider(bucket_, b, node);
-            visited->Reset();
-            auto candidates =
-                searcher.Search(graph, distance_provider, visited, search_param, nullptr, nullptr);
-            mutually_connect_new_element(
-                node, candidates, graph, distance_provider, mutexes, allocator_);
-        }
-
-        bucket_graphs_[b] = std::move(graph);
-    };
-
-    if (this->thread_pool_ != nullptr) {
-        std::vector<std::future<void>> futures;
-        futures.reserve(bucket_count);
-        for (BucketIdType b = 0; b < bucket_count; ++b) {
-            futures.emplace_back(this->thread_pool_->GeneralEnqueue(build_one_bucket, b));
-        }
-        for (auto& future : futures) {
-            future.get();
-        }
-    } else {
-        for (BucketIdType b = 0; b < bucket_count; ++b) {
-            build_one_bucket(b);
-        }
-    }
-}
-DatasetPtr
-IVF::KnnSearch(const DatasetPtr& query,
-               int64_t k,
-               const std::string& parameters,
-               const FilterPtr& filter) const {
-    SearchRequest req;
-    req.mode_ = SearchMode::KNN_SEARCH;
-    req.query_ = query;
-    req.topk_ = k;
-    req.params_str_ = parameters;
-    req.threshold_ = ParseSearchThreshold(parameters);
-    if (filter != nullptr) {
-        req.filter_ = filter;
-    }
-    return this->SearchWithRequest(req);
-}
-
-DatasetPtr
-IVF::RangeSearch(const DatasetPtr& query,
-                 float radius,
-                 const std::string& parameters,
-                 const FilterPtr& filter,
-                 int64_t limited_size) const {
-    SearchRequest req;
-    req.mode_ = SearchMode::RANGE_SEARCH;
-    req.query_ = query;
-    req.radius_ = radius;
-    req.limited_size_ = limited_size;
-    req.params_str_ = parameters;
-    if (filter != nullptr) {
-        req.filter_ = filter;
-    }
-    return this->SearchWithRequest(req);
-}
-
 int64_t
 IVF::GetNumElements() const {
     return this->total_elements_ - this->delete_count_;
@@ -780,726 +315,6 @@ IVF::get_location(InnerIdType inner_id) const {
     return {bucket_id, offset_id};
 }
 
-uint32_t
-IVF::Remove(const std::vector<int64_t>& ids, RemoveMode mode) {
-    uint32_t delete_count = 0;
-    if (mode == RemoveMode::MARK_REMOVE) {
-        std::scoped_lock label_lock(this->label_lookup_mutex_);
-        delete_count = this->label_table_->MarkRemove(ids);
-        delete_count_ += delete_count;
-    }
-    return delete_count;
-}
-
-void
-IVF::UpdateAttribute(int64_t id, const AttributeSet& new_attrs) {
-    auto inner_id = this->label_table_->GetIdByLabel(id);
-    auto [bucket_id, offset_id] = this->get_location(inner_id);
-    this->attr_filter_index_->UpdateBitsetsByAttr(new_attrs, offset_id, bucket_id);
-}
-
-void
-IVF::UpdateAttribute(int64_t id, const AttributeSet& new_attrs, const AttributeSet& origin_attrs) {
-    auto inner_id = this->label_table_->GetIdByLabel(id);
-    auto [bucket_id, offset_id] = this->get_location(inner_id);
-    this->attr_filter_index_->UpdateBitsetsByAttr(new_attrs, offset_id, bucket_id, origin_attrs);
-}
-
-#define WRITE_DATACELL_WITH_NAME(writer, name, datacell)            \
-    datacell_offsets[(name)].SetInt(offset);                        \
-    auto datacell##_start = (writer).GetCursor();                   \
-    (datacell)->Serialize(writer);                                  \
-    auto datacell##_size = (writer).GetCursor() - datacell##_start; \
-    datacell_sizes[(name)].SetInt(datacell##_size);                 \
-    offset += datacell##_size;
-
-void
-IVF::Serialize(StreamWriter& writer) const {
-    JsonType datacell_offsets;
-    JsonType datacell_sizes;
-    uint64_t offset = 0;
-
-    WRITE_DATACELL_WITH_NAME(writer, "bucket", bucket_);
-    WRITE_DATACELL_WITH_NAME(writer, "partition_strategy", partition_strategy_);
-    WRITE_DATACELL_WITH_NAME(writer, "label_table", label_table_);
-
-    if (use_reorder_) {
-        if (precise_bucket_ != nullptr) {
-            WRITE_DATACELL_WITH_NAME(writer, "precise_bucket", precise_bucket_);
-        } else {
-            WRITE_DATACELL_WITH_NAME(writer, "reorder_codes", reorder_codes_);
-        }
-    }
-
-    if (use_attribute_filter_) {
-        WRITE_DATACELL_WITH_NAME(writer, "attr_filter_index", attr_filter_index_);
-    }
-
-    if (graph_build_threshold_ > 0 && graph_param_ != nullptr &&
-        has_any_fresh_bucket_graph(bucket_, bucket_graphs_)) {
-        datacell_offsets["bucket_graphs"].SetInt(offset);
-        auto bucket_graphs_start = writer.GetCursor();
-        StreamWriter::WriteObj(writer, graph_build_threshold_);
-        StreamWriter::WriteString(writer, serialize_ivf_graph_param(graph_param_).Dump());
-        int64_t graph_count = 0;
-        for (BucketIdType b = 0; b < static_cast<BucketIdType>(bucket_graphs_.size()); ++b) {
-            if (bucket_graphs_[b] != nullptr &&
-                bucket_graphs_[b]->TotalCount() == bucket_->GetBucketSize(b)) {
-                ++graph_count;
-            }
-        }
-        StreamWriter::WriteObj(writer, graph_count);
-        for (BucketIdType b = 0; b < static_cast<BucketIdType>(bucket_graphs_.size()); ++b) {
-            if (bucket_graphs_[b] != nullptr &&
-                bucket_graphs_[b]->TotalCount() == bucket_->GetBucketSize(b)) {
-                StreamWriter::WriteObj(writer, b);
-                bucket_graphs_[b]->Serialize(writer);
-            }
-        }
-        auto bucket_graphs_size = writer.GetCursor() - bucket_graphs_start;
-        datacell_sizes["bucket_graphs"].SetInt(bucket_graphs_size);
-        offset += bucket_graphs_size;
-    }
-
-    // serialize footer (introduced since v0.15)
-    JsonType basic_info;
-    basic_info["total_elements"].SetInt(this->total_elements_);
-    basic_info["use_reorder"].SetBool(this->use_reorder_);
-    basic_info["is_trained"].SetBool(this->is_trained_);
-    basic_info[DIM].SetInt(this->dim_);
-    basic_info[EXTRA_INFO_SIZE].SetInt(0);
-    basic_info[INDEX_PARAM].SetString(this->create_param_ptr_->ToString());
-    basic_info["data_type"].SetInt(static_cast<int64_t>(this->data_type_));
-    basic_info["metric"].SetInt(static_cast<int64_t>(this->metric_));
-
-    auto metadata = std::make_shared<Metadata>();
-    metadata->Set(BASIC_INFO, basic_info);
-    metadata->Set("datacell_offsets", datacell_offsets);
-    metadata->Set("datacell_sizes", datacell_sizes);
-
-    auto footer = std::make_shared<Footer>(metadata);
-    footer->Write(writer);
-}
-
-MetadataPtr
-IVF::collect_streaming_header() const {
-    auto metadata = std::make_shared<Metadata>();
-    metadata->Set("format", "vsag_stream_v1");
-    metadata->Set("index_name", this->GetName());
-
-    JsonType basic_info;
-    basic_info["total_elements"].SetInt(this->total_elements_);
-    basic_info["use_reorder"].SetBool(this->use_reorder_);
-    basic_info["is_trained"].SetBool(this->is_trained_);
-    basic_info[DIM].SetInt(this->dim_);
-    basic_info[EXTRA_INFO_SIZE].SetInt(0);
-    basic_info[INDEX_PARAM].SetString(this->create_param_ptr_->ToString());
-    basic_info["data_type"].SetInt(static_cast<int64_t>(this->data_type_));
-    basic_info["metric"].SetInt(static_cast<int64_t>(this->metric_));
-    metadata->Set(BASIC_INFO, basic_info);
-
-    JsonType manifest;
-    auto bucket_tag = static_cast<uint32_t>(StreamSerializationTag::IVF_BUCKET);
-    auto partition_tag = static_cast<uint32_t>(StreamSerializationTag::IVF_PARTITION_STRATEGY);
-    auto label_tag = static_cast<uint32_t>(StreamSerializationTag::LABEL_TABLE);
-    AppendStreamingManifestBlock(manifest,
-                                 bucket_tag,
-                                 StreamSerializationBlockCurrentVersion(bucket_tag),
-                                 StreamSerializationTagCritical(bucket_tag));
-    AppendStreamingManifestBlock(manifest,
-                                 partition_tag,
-                                 StreamSerializationBlockCurrentVersion(partition_tag),
-                                 StreamSerializationTagCritical(partition_tag));
-    AppendStreamingManifestBlock(manifest,
-                                 label_tag,
-                                 StreamSerializationBlockCurrentVersion(label_tag),
-                                 StreamSerializationTagCritical(label_tag));
-    if (this->use_reorder_) {
-        auto tag = static_cast<uint32_t>(precise_bucket_ != nullptr
-                                             ? StreamSerializationTag::IVF_PRECISE_BUCKET
-                                             : StreamSerializationTag::HIGH_PRECISION_CODES);
-        AppendStreamingManifestBlock(manifest,
-                                     tag,
-                                     StreamSerializationBlockCurrentVersion(tag),
-                                     StreamSerializationTagCritical(tag));
-    }
-    if (this->use_attribute_filter_) {
-        auto tag = static_cast<uint32_t>(StreamSerializationTag::ATTRIBUTE_FILTER);
-        AppendStreamingManifestBlock(manifest,
-                                     tag,
-                                     StreamSerializationBlockCurrentVersion(tag),
-                                     StreamSerializationTagCritical(tag));
-    }
-    if (graph_build_threshold_ > 0 && graph_param_ != nullptr &&
-        has_any_fresh_bucket_graph(bucket_, bucket_graphs_)) {
-        auto tag = static_cast<uint32_t>(StreamSerializationTag::IVF_BUCKET_GRAPH);
-        AppendStreamingManifestBlock(manifest,
-                                     tag,
-                                     StreamSerializationBlockCurrentVersion(tag),
-                                     StreamSerializationTagCritical(tag));
-    }
-    metadata->Set("block_manifest", manifest);
-    metadata->SetEmptyIndex(this->GetNumElements() == 0);
-    return metadata;
-}
-
-void
-IVF::serialize_streaming_body(StreamWriter& writer) const {
-    auto bucket_tag = static_cast<uint32_t>(StreamSerializationTag::IVF_BUCKET);
-    auto partition_tag = static_cast<uint32_t>(StreamSerializationTag::IVF_PARTITION_STRATEGY);
-    auto label_tag = static_cast<uint32_t>(StreamSerializationTag::LABEL_TABLE);
-
-    WriteStreamingBlock(
-        writer, bucket_tag, StreamSerializationTagCritical(bucket_tag), [this](StreamWriter& w) {
-            this->bucket_->Serialize(w);
-        });
-    WriteStreamingBlock(writer,
-                        partition_tag,
-                        StreamSerializationTagCritical(partition_tag),
-                        [this](StreamWriter& w) { this->partition_strategy_->Serialize(w); });
-    WriteStreamingBlock(
-        writer, label_tag, StreamSerializationTagCritical(label_tag), [this](StreamWriter& w) {
-            this->label_table_->Serialize(w);
-        });
-    if (this->use_reorder_) {
-        auto tag = static_cast<uint32_t>(precise_bucket_ != nullptr
-                                             ? StreamSerializationTag::IVF_PRECISE_BUCKET
-                                             : StreamSerializationTag::HIGH_PRECISION_CODES);
-        WriteStreamingBlock(
-            writer, tag, StreamSerializationTagCritical(tag), [this](StreamWriter& w) {
-                if (this->precise_bucket_ != nullptr) {
-                    this->precise_bucket_->Serialize(w);
-                } else {
-                    this->reorder_codes_->Serialize(w);
-                }
-            });
-    }
-    if (this->use_attribute_filter_) {
-        auto tag = static_cast<uint32_t>(StreamSerializationTag::ATTRIBUTE_FILTER);
-        WriteStreamingBlock(
-            writer, tag, StreamSerializationTagCritical(tag), [this](StreamWriter& w) {
-                this->attr_filter_index_->Serialize(w);
-            });
-    }
-    if (graph_build_threshold_ > 0 && graph_param_ != nullptr &&
-        has_any_fresh_bucket_graph(bucket_, bucket_graphs_)) {
-        auto tag = static_cast<uint32_t>(StreamSerializationTag::IVF_BUCKET_GRAPH);
-        WriteStreamingBlock(
-            writer, tag, StreamSerializationTagCritical(tag), [this](StreamWriter& w) {
-                StreamWriter::WriteObj(w, graph_build_threshold_);
-                StreamWriter::WriteString(w, serialize_ivf_graph_param(graph_param_).Dump());
-                auto graph_count = static_cast<uint64_t>(bucket_graphs_.size());
-                StreamWriter::WriteObj(w, graph_count);
-                for (uint64_t i = 0; i < graph_count; ++i) {
-                    bool has_graph = (bucket_graphs_[i] != nullptr &&
-                                      bucket_graphs_[i]->TotalCount() ==
-                                          bucket_->GetBucketSize(static_cast<BucketIdType>(i)));
-                    StreamWriter::WriteObj(w, has_graph);
-                    if (has_graph) {
-                        StreamWriter::WriteObj(w, static_cast<BucketIdType>(i));
-                        bucket_graphs_[i]->Serialize(w);
-                    }
-                }
-            });
-    }
-}
-
-void
-IVF::deserialize_streaming_body(StreamReader& reader, const MetadataPtr& metadata) {
-    this->read_streaming_body(reader, metadata);
-}
-
-void
-IVF::load_streaming_body(StreamReader& reader,
-                         const MetadataPtr& metadata,
-                         const LoadParameters& parameters) {
-    this->read_streaming_body(reader, metadata, &parameters);
-}
-
-void
-IVF::read_streaming_body(StreamReader& reader,
-                         const MetadataPtr& metadata,
-                         const LoadParameters* load_parameters) {
-    auto basic_info = metadata->Get(BASIC_INFO);
-    this->total_elements_ = basic_info["total_elements"].GetInt();
-    this->use_reorder_ = basic_info["use_reorder"].GetBool();
-    this->is_trained_ = basic_info["is_trained"].GetBool();
-    if (precise_bucket_ != nullptr and not basic_info.Contains(INDEX_PARAM)) {
-        throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                            "IVF precise bucket requires persisted index parameters");
-    }
-    if (basic_info.Contains(INDEX_PARAM)) {
-        auto index_param = std::make_shared<IVFParameter>();
-        index_param->FromString(basic_info[INDEX_PARAM].GetString());
-        if (not this->create_param_ptr_->CheckCompatibility(index_param)) {
-            auto message = fmt::format("IVF index parameter not match, current: {}, new: {}",
-                                       this->create_param_ptr_->ToString(),
-                                       index_param->ToString());
-            logger::error(message);
-            throw VsagException(ErrorType::INVALID_ARGUMENT, message);
-        }
-    }
-
-    bool loaded_bucket = false;
-    bool loaded_partition = false;
-    bool loaded_label_table = false;
-    bool loaded_precise_codes = false;
-    bool loaded_attribute_filter = false;
-
-    ReaderIOParamPtr precise_reader_param = nullptr;
-    auto ivf_param = std::dynamic_pointer_cast<IVFParameter>(create_param_ptr_);
-    if (ivf_param != nullptr && ivf_param->precise_codes_param != nullptr &&
-        ivf_param->precise_codes_param->io_parameter != nullptr &&
-        ivf_param->precise_codes_param->io_parameter->GetTypeName() == IO_TYPE_VALUE_READER_IO) {
-        constexpr const char* precise_reader_key = "precise_reader";
-        if (load_parameters == nullptr or not load_parameters->HasReader(precise_reader_key)) {
-            throw VsagException(ErrorType::INVALID_ARGUMENT,
-                                "reader-backed IVF precise codes require precise_reader");
-        }
-        auto precise_reader = load_parameters->GetReader(precise_reader_key);
-        if (precise_reader == nullptr) {
-            throw VsagException(ErrorType::INVALID_ARGUMENT, "precise_reader is null");
-        }
-        precise_reader_param = std::dynamic_pointer_cast<ReaderIOParameter>(
-            ivf_param->precise_codes_param->io_parameter);
-        if (precise_reader_param == nullptr) {
-            throw VsagException(ErrorType::INTERNAL_ERROR,
-                                "IVF precise reader IO parameter is invalid");
-        }
-        precise_reader_param->reader = std::move(precise_reader);
-    }
-
-    while (true) {
-        auto block_header = StreamBlockHeader::Read(reader);
-        if (block_header.IsSectionEnd()) {
-            break;
-        }
-        BoundedForwardReader block_reader(&reader, block_header.value_len);
-        if (!StreamSerializationBlockVersionSupported(block_header.tag,
-                                                      block_header.block_version)) {
-            if (block_header.IsCritical()) {
-                throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                                    fmt::format("unsupported IVF streaming block version: tag={}, "
-                                                "name={}, version={}, flags={}, value_len={}",
-                                                block_header.tag,
-                                                StreamSerializationTagName(block_header.tag),
-                                                block_header.block_version,
-                                                block_header.flags,
-                                                block_header.value_len));
-            }
-            block_reader.SkipRemaining();
-            continue;
-        }
-
-        auto read_precise_block = [&](const auto& deserialize, const auto& init_io) {
-            if (precise_reader_param == nullptr) {
-                ReadSeekableBlockPayload(block_reader, block_header, deserialize);
-                return;
-            }
-            block_reader.SkipRemaining();
-            ReadExternalBlockPayload(precise_reader_param->reader, block_header, deserialize);
-            init_io(precise_reader_param);
-        };
-
-        switch (static_cast<StreamSerializationTag>(block_header.tag)) {
-            case StreamSerializationTag::IVF_BUCKET:
-                ReadSeekableBlockPayload(block_reader, block_header, [this](StreamReader& block) {
-                    this->bucket_->Deserialize(block);
-                });
-                loaded_bucket = true;
-                break;
-            case StreamSerializationTag::IVF_PARTITION_STRATEGY:
-                ReadSeekableBlockPayload(block_reader, block_header, [this](StreamReader& block) {
-                    this->partition_strategy_->Deserialize(block);
-                });
-                loaded_partition = true;
-                break;
-            case StreamSerializationTag::LABEL_TABLE:
-                ReadSeekableBlockPayload(block_reader, block_header, [this](StreamReader& block) {
-                    this->label_table_->Deserialize(block);
-                });
-                loaded_label_table = true;
-                break;
-            case StreamSerializationTag::HIGH_PRECISION_CODES:
-                if (this->use_reorder_ and this->reorder_codes_ != nullptr) {
-                    read_precise_block(
-                        [this](StreamReader& block) { this->reorder_codes_->Deserialize(block); },
-                        [this](const IOParamPtr& io_param) {
-                            this->reorder_codes_->InitIO(io_param);
-                        });
-                    loaded_precise_codes = true;
-                }
-                break;
-            case StreamSerializationTag::IVF_PRECISE_BUCKET:
-                if (this->use_reorder_ and this->precise_bucket_ != nullptr) {
-                    read_precise_block(
-                        [this](StreamReader& block) { this->precise_bucket_->Deserialize(block); },
-                        [this](const IOParamPtr& io_param) {
-                            this->precise_bucket_->InitIO(io_param);
-                        });
-                    loaded_precise_codes = true;
-                }
-                break;
-            case StreamSerializationTag::ATTRIBUTE_FILTER:
-                if (this->use_attribute_filter_) {
-                    ReadSeekableBlockPayload(
-                        block_reader, block_header, [this](StreamReader& block) {
-                            this->attr_filter_index_->Deserialize(block);
-                        });
-                    loaded_attribute_filter = true;
-                    this->has_attribute_ = true;
-                }
-                break;
-            case StreamSerializationTag::IVF_BUCKET_GRAPH: {
-                ReadSeekableBlockPayload(block_reader, block_header, [this](StreamReader& block) {
-                    StreamReader::ReadObj(block, this->graph_build_threshold_);
-                    this->graph_param_ = get_ivf_graph_param(StreamReader::ReadString(block));
-                    uint64_t graph_count = 0;
-                    StreamReader::ReadObj(block, graph_count);
-                    auto bucket_count = this->bucket_->bucket_count_;
-                    if (graph_count > static_cast<uint64_t>(bucket_count)) {
-                        throw VsagException(
-                            ErrorType::INVALID_BINARY,
-                            fmt::format("bucket graph_count {} exceeds bucket_count {}",
-                                        graph_count,
-                                        bucket_count));
-                    }
-                    this->bucket_graphs_.resize(bucket_count);
-                    for (uint64_t i = 0; i < graph_count; ++i) {
-                        bool has_graph = false;
-                        StreamReader::ReadObj(block, has_graph);
-                        if (has_graph) {
-                            BucketIdType bid = 0;
-                            StreamReader::ReadObj(block, bid);
-                            auto graph = GraphInterface::MakeInstance(this->graph_param_,
-                                                                      this->common_param_);
-                            graph->Deserialize(block);
-                            if (bid >= 0 && bid < static_cast<BucketIdType>(bucket_count)) {
-                                const auto total = graph->TotalCount();
-                                const auto expected_total = bucket_->GetBucketSize(bid);
-                                if (total != expected_total || total > graph->MaxCapacity()) {
-                                    throw VsagException(
-                                        ErrorType::INVALID_BINARY,
-                                        "corrupt bucket graph: invalid total count");
-                                }
-                                for (InnerIdType nid = 0; nid < total; ++nid) {
-                                    if (graph->GetNeighborSize(nid) > graph->MaximumDegree()) {
-                                        throw VsagException(ErrorType::INVALID_BINARY,
-                                                            "corrupt bucket graph: neighbor count "
-                                                            "exceeds maximum degree");
-                                    }
-                                    Vector<InnerIdType> nbrs(this->allocator_);
-                                    graph->GetNeighbors(nid, nbrs);
-                                    for (auto nb : nbrs) {
-                                        if (nb >= total) {
-                                            throw VsagException(
-                                                ErrorType::INVALID_BINARY,
-                                                "corrupt bucket graph: neighbor ID out of range");
-                                        }
-                                    }
-                                }
-                                this->bucket_graphs_[bid] = std::move(graph);
-                            }
-                        }
-                    }
-                });
-                if (graph_build_threshold_ > 0 && this->bucket_searcher_ != nullptr) {
-                    // Replace flat searcher with graph searcher now that graphs are loaded.
-                    this->bucket_searcher_ = std::make_shared<GraphBucketSearcher>(
-                        graph_build_threshold_, bucket_graphs_, allocator_);
-                }
-                break;
-            }
-            default:
-                if (block_header.IsCritical()) {
-                    throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                                        fmt::format("unknown IVF streaming serialization block: "
-                                                    "tag={}, name={}, version={}, flags={}, "
-                                                    "value_len={}",
-                                                    block_header.tag,
-                                                    StreamSerializationTagName(block_header.tag),
-                                                    block_header.block_version,
-                                                    block_header.flags,
-                                                    block_header.value_len));
-                }
-                break;
-        }
-        block_reader.SkipRemaining();
-    }
-
-    if (!loaded_bucket || !loaded_partition || !loaded_label_table) {
-        throw VsagException(ErrorType::READ_ERROR,
-                            "IVF streaming serialization required block is missing");
-    }
-    if (this->use_reorder_ && !loaded_precise_codes) {
-        throw VsagException(ErrorType::READ_ERROR,
-                            "IVF streaming serialization reorder block is missing");
-    }
-    if (this->use_attribute_filter_ && !loaded_attribute_filter) {
-        throw VsagException(ErrorType::READ_ERROR,
-                            "IVF streaming serialization attribute filter block is missing");
-    }
-    if (this->bucket_->GetQuantizerName() == QUANTIZATION_TYPE_VALUE_FP32) {
-        this->has_raw_vector_ = true;
-    }
-    this->fill_location_map();
-    this->cal_memory_usage();
-}
-
-#define READ_DATACELL_WITH_NAME(reader, name, datacell)                       \
-    reader.PushSeek(datacell_offsets[(name)].GetInt());                       \
-    (datacell)->Deserialize((reader).Slice(datacell_sizes[(name)].GetInt())); \
-    (reader).PopSeek();
-
-void
-IVF::Deserialize(StreamReader& reader) {
-    // try to deserialize footer (only in new version)
-    auto footer = Footer::Parse(reader);
-
-    BufferStreamReader buffer_reader(
-        &reader, std::numeric_limits<uint64_t>::max(), this->allocator_);
-
-    if (footer == nullptr) {  // old format, DON'T EDIT, remove in the future
-        if (precise_bucket_ != nullptr) {
-            throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                                "legacy IVF serialization does not support precise bucket");
-        }
-        logger::debug("parse with v0.14 version format");
-
-        StreamReader::ReadObj(buffer_reader, this->total_elements_);
-        StreamReader::ReadObj(buffer_reader, this->use_reorder_);
-        StreamReader::ReadObj(buffer_reader, this->is_trained_);
-
-        this->bucket_->Deserialize(buffer_reader);
-        this->partition_strategy_->Deserialize(buffer_reader);
-        this->label_table_->Deserialize(buffer_reader);
-        if (use_reorder_) {
-            this->reorder_codes_->Deserialize(buffer_reader);
-        }
-
-        if (use_attribute_filter_) {
-            this->attr_filter_index_->Deserialize(buffer_reader);
-            this->has_attribute_ = true;
-        }
-    } else {  // create like `else if ( ver in [v0.15, v0.17] )` here if need in the future
-        logger::debug("parse with new version format");
-
-        auto metadata = footer->GetMetadata();
-        if (metadata->EmptyIndex()) {
-            return;
-        }
-
-        auto basic_info = metadata->Get(BASIC_INFO);
-        this->total_elements_ = basic_info["total_elements"].GetInt();
-        this->use_reorder_ = basic_info["use_reorder"].GetBool();
-        this->is_trained_ = basic_info["is_trained"].GetBool();
-        if (precise_bucket_ != nullptr and not basic_info.Contains(INDEX_PARAM)) {
-            throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                                "IVF precise bucket requires persisted index parameters");
-        }
-        if (basic_info.Contains(INDEX_PARAM)) {
-            auto param_str = basic_info[INDEX_PARAM].GetString();
-            auto index_param = std::make_shared<IVFParameter>();
-            index_param->FromString(param_str);
-            if (not this->create_param_ptr_->CheckCompatibility(index_param)) {
-                auto message = fmt::format("IVF index parameter not match, current: {}, new: {}",
-                                           this->create_param_ptr_->ToString(),
-                                           index_param->ToString());
-                logger::error(message);
-                throw VsagException(ErrorType::INVALID_ARGUMENT, message);
-            }
-        }
-
-        JsonType datacell_offsets = metadata->Get(DATACELL_OFFSETS);
-        logger::debug("datacell_offsets: {}", datacell_offsets.Dump());
-        JsonType datacell_sizes = metadata->Get(DATACELL_SIZES);
-        logger::debug("datacell_sizes: {}", datacell_sizes.Dump());
-
-        READ_DATACELL_WITH_NAME(buffer_reader, "bucket", this->bucket_);
-        READ_DATACELL_WITH_NAME(buffer_reader, "partition_strategy", this->partition_strategy_);
-        READ_DATACELL_WITH_NAME(buffer_reader, "label_table", this->label_table_);
-        if (use_reorder_) {
-            if (precise_bucket_ != nullptr) {
-                READ_DATACELL_WITH_NAME(buffer_reader, "precise_bucket", this->precise_bucket_);
-            } else {
-                READ_DATACELL_WITH_NAME(buffer_reader, "reorder_codes", this->reorder_codes_);
-            }
-        }
-        if (use_attribute_filter_) {
-            READ_DATACELL_WITH_NAME(buffer_reader, "attr_filter_index", this->attr_filter_index_);
-            this->has_attribute_ = true;
-        }
-        if (this->bucket_->GetQuantizerName() == QUANTIZATION_TYPE_VALUE_FP32) {
-            this->has_raw_vector_ = true;
-        }
-
-        if (datacell_offsets.Contains("bucket_graphs")) {
-            auto bucket_count = bucket_->bucket_count_;
-            bucket_graphs_.resize(bucket_count);
-            buffer_reader.PushSeek(datacell_offsets["bucket_graphs"].GetInt());
-            auto graph_reader = buffer_reader.Slice(datacell_sizes["bucket_graphs"].GetInt());
-            int64_t stored_threshold = 0;
-            StreamReader::ReadObj(graph_reader, stored_threshold);
-            graph_build_threshold_ = stored_threshold;
-            graph_param_ = get_ivf_graph_param(StreamReader::ReadString(graph_reader));
-            int64_t graph_count = 0;
-            StreamReader::ReadObj(graph_reader, graph_count);
-            CHECK_ARGUMENT(graph_count >= 0, "bucket graph_count is negative");
-            CHECK_ARGUMENT(graph_count <= bucket_count, "bucket graph_count exceeds bucket_count");
-            for (int64_t g = 0; g < graph_count; ++g) {
-                BucketIdType bid = 0;
-                StreamReader::ReadObj(graph_reader, bid);
-                auto graph = GraphInterface::MakeInstance(graph_param_, this->common_param_);
-                graph->Deserialize(graph_reader);
-                if (bid >= 0 && bid < bucket_count) {
-                    const auto total = graph->TotalCount();
-                    const auto expected_total = bucket_->GetBucketSize(bid);
-                    if (total != expected_total || total > graph->MaxCapacity()) {
-                        throw VsagException(ErrorType::INVALID_BINARY,
-                                            "corrupt bucket graph: invalid total count");
-                    }
-                    for (InnerIdType nid = 0; nid < total; ++nid) {
-                        if (graph->GetNeighborSize(nid) > graph->MaximumDegree()) {
-                            throw VsagException(
-                                ErrorType::INVALID_BINARY,
-                                "corrupt bucket graph: neighbor count exceeds maximum degree");
-                        }
-                        Vector<InnerIdType> nbrs(allocator_);
-                        graph->GetNeighbors(nid, nbrs);
-                        for (auto nb : nbrs) {
-                            if (nb >= total) {
-                                throw VsagException(
-                                    ErrorType::INVALID_BINARY,
-                                    "corrupt bucket graph: neighbor ID out of range");
-                            }
-                        }
-                    }
-                    bucket_graphs_[bid] = std::move(graph);
-                }
-            }
-            buffer_reader.PopSeek();
-            if (graph_build_threshold_ > 0) {
-                this->bucket_searcher_ = std::make_shared<GraphBucketSearcher>(
-                    graph_build_threshold_, bucket_graphs_, allocator_);
-            }
-        }
-    }
-    this->fill_location_map();
-    this->cal_memory_usage();
-}
-
-InnerSearchParam
-IVF::create_search_param(const std::string& parameters, const FilterPtr& filter) const {
-    InnerSearchParam param;
-    param.is_inner_id_allowed = this->create_search_filter(filter);
-    auto search_param = IVFSearchParameters::FromJson(parameters);
-    if (search_param.disable_bucket_scan) {
-        param.scan_bucket_size = static_cast<BucketIdType>(search_param.scan_buckets_count);
-    } else {
-        param.scan_bucket_size = std::min(
-            static_cast<BucketIdType>(search_param.scan_buckets_count), bucket_->bucket_count_);
-    }
-    param.disable_bucket_scan = search_param.disable_bucket_scan;
-    param.factor = search_param.topk_factor;
-    param.enable_reorder = search_param.enable_reorder;
-    param.first_order_scan_ratio = search_param.first_order_scan_ratio;
-    param.parallel_search_thread_count = search_param.parallel_search_thread_count;
-    param.ef = static_cast<uint64_t>(search_param.ef_search);
-    if (search_param.enable_time_record) {
-        param.time_cost = std::make_shared<Timer>();
-        param.time_cost->SetThreshold(search_param.timeout_ms);
-    }
-    return param;
-}
-
-DatasetPtr
-IVF::route_buckets_only(const DatasetPtr& query,
-                        const InnerSearchParam& param,
-                        QueryContext& ctx) const {
-    const auto num_queries = query->GetNumElements();
-    const auto* query_data = query->GetFloat32Vectors();
-    const auto buckets_per_query = param.scan_bucket_size;
-    const auto candidate_buckets =
-        partition_strategy_->ClassifyDatasForSearch(query_data, num_queries, param, &ctx);
-
-    auto result = Dataset::Make();
-    if (num_queries == 0 || buckets_per_query == 0) {
-        return result->NumElements(0)->Dim(0);
-    }
-
-    auto* alloc = (ctx.alloc != nullptr) ? ctx.alloc : allocator_;
-    const auto total_slots = num_queries * buckets_per_query;
-    auto* ids = static_cast<int64_t*>(alloc->Allocate(sizeof(int64_t) * total_slots));
-    auto* distances = static_cast<float*>(alloc->Allocate(sizeof(float) * total_slots));
-    const auto dim = partition_strategy_->dim_;
-    const auto metric = partition_strategy_->metric_type_;
-
-    Vector<float> centroid(dim, allocator_);
-    Vector<float> norm_query(dim, allocator_);
-    Vector<float> norm_centroid(dim, allocator_);
-    for (int64_t q = 0; q < num_queries; ++q) {
-        const auto* query_vec = query_data + q * dim;
-        if (metric == MetricType::METRIC_TYPE_COSINE) {
-            Normalize(query_vec, norm_query.data(), dim);
-        }
-        for (int64_t b = 0; b < buckets_per_query; ++b) {
-            const auto idx = q * buckets_per_query + b;
-            const auto bucket_id = candidate_buckets[idx];
-            if (bucket_id == INVALID_BUCKET_ID) {
-                ids[idx] = -1;
-                distances[idx] = std::numeric_limits<float>::infinity();
-                continue;
-            }
-            partition_strategy_->GetCentroid(bucket_id, centroid);
-            float dist = 0.0F;
-            if (metric == MetricType::METRIC_TYPE_L2SQR) {
-                for (int64_t d = 0; d < dim; ++d) {
-                    auto diff = query_vec[d] - centroid[d];
-                    dist += diff * diff;
-                }
-            } else if (metric == MetricType::METRIC_TYPE_COSINE) {
-                Normalize(centroid.data(), norm_centroid.data(), dim);
-                for (int64_t d = 0; d < dim; ++d) {
-                    dist += norm_query[d] * norm_centroid[d];
-                }
-                dist = 1.0F - dist;
-            } else {
-                for (int64_t d = 0; d < dim; ++d) {
-                    dist += query_vec[d] * centroid[d];
-                }
-                dist = 1.0F - dist;
-            }
-            ids[idx] = static_cast<int64_t>(bucket_id);
-            distances[idx] = dist;
-            if (ctx.stats != nullptr) {
-                ctx.stats->AddDistance(SearchStatistics::DistancePhase::ROUTING,
-                                       DistanceEvaluationBackend::FP32);
-            }
-        }
-    }
-
-    return result->NumElements(num_queries)
-        ->Dim(buckets_per_query)
-        ->Ids(ids)
-        ->Distances(distances)
-        ->Owner(true, alloc);
-}
-
-DatasetPtr
-IVF::reorder(int64_t topk,
-             DistHeapPtr& input,
-             const float* query,
-             const InnerSearchParam& param,
-             QueryContext& ctx,
-             ReasoningContext* reasoning_ctx,
-             const std::optional<float>& distance_threshold) const {
-    auto reorder_heap =
-        reorder_->Reorder(input, query, topk, ctx, nullptr, nullptr, distance_threshold);
-    auto dataset_results = this->pack_knn_result(reorder_heap, ctx.alloc);
-
-    return dataset_results;
-}
-
 InnerIndexPtr
 IVF::ExportModel(const IndexCommonParam& param) const {
     auto index = std::make_shared<IVF>(this->create_param_ptr_, param);
@@ -1514,301 +329,6 @@ IVF::ExportModel(const IndexCommonParam& param) const {
     }
     index->is_trained_ = this->is_trained_;
     return index;
-}
-
-template <InnerSearchMode mode>
-DistHeapPtr
-IVF::search(const DatasetPtr& query,
-            const InnerSearchParam& param,
-            QueryContext& ctx,
-            ReasoningContext* reasoning_ctx) const {
-    const auto* query_data = query->GetFloat32Vectors();
-    Vector<float> normalize_data(dim_, allocator_);
-    Vector<BucketIdType> candidate_buckets(allocator_);
-    if (not param.bucket_ids.empty()) {
-        candidate_buckets.reserve(param.bucket_ids.size());
-        for (auto id : param.bucket_ids) {
-            candidate_buckets.push_back(static_cast<BucketIdType>(id));
-        }
-    } else {
-        candidate_buckets = partition_strategy_->ClassifyDatasForSearch(query_data, 1, param, &ctx);
-    }
-    if (reasoning_ctx != nullptr) {
-        reasoning_ctx->RecordBucketSelection(candidate_buckets);
-    }
-    auto computer = bucket_->FactoryComputer(query_data);
-
-    int64_t topk = param.topk;
-    if constexpr (mode == RANGE_SEARCH) {
-        topk = param.range_search_limit_size;
-        if (topk < 0) {
-            topk = this->GetNumElements();
-        }
-    }
-    // Scale topk to ensure sufficient candidates after deduplication when buckets_per_data_ > 1
-    int64_t origin_topk = topk;
-    if (buckets_per_data_ > 1) {
-        if (topk <= std::numeric_limits<int64_t>::max() / buckets_per_data_) {
-            topk *= buckets_per_data_;
-        } else {
-            topk = std::numeric_limits<int64_t>::max();
-        }
-    }
-
-    DistHeapPtr search_result = nullptr;
-
-    auto bucket_count = candidate_buckets.size();
-    auto search_thread_count = param.parallel_search_thread_count;
-    if (this->thread_pool_ == nullptr) {
-        search_thread_count = 1;
-    }
-    std::vector<DistHeapPtr> heaps(search_thread_count);
-    std::atomic<uint64_t> cur_bucket_num(0);
-    auto search_func = [&](int64_t thread_id) -> void {
-        heaps[thread_id] = DistanceHeap::MakeInstanceBySize<true, false>(this->allocator_, topk);
-        auto& heap = heaps[thread_id];
-        Vector<float> dist(allocator_);
-        uint64_t i = cur_bucket_num.fetch_add(1);
-        for (; i < bucket_count; i = cur_bucket_num.fetch_add(1)) {
-            if (param.time_cost != nullptr and param.time_cost->CheckOvertime() and
-                ctx.stats != nullptr) {
-                ctx.stats->is_timeout.store(true, std::memory_order_relaxed);
-                break;
-            }
-            auto bucket_id = candidate_buckets[i];
-            if (bucket_id == INVALID_BUCKET_ID) {
-                break;
-            }
-            bucket_searcher_->Search(bucket_id,
-                                     bucket_,
-                                     computer,
-                                     param,
-                                     thread_id,
-                                     topk,
-                                     buckets_per_data_,
-                                     heap,
-                                     dist,
-                                     reasoning_ctx);
-        }
-    };
-    std::vector<std::future<void>> futures;
-    if (this->thread_pool_ != nullptr and search_thread_count > 1) {
-        for (int64_t thread_id = 0; thread_id < search_thread_count; ++thread_id) {
-            auto future = this->thread_pool_->GeneralEnqueue(search_func, thread_id);
-            futures.emplace_back(std::move(future));
-        }
-    } else {
-        search_func(0);
-        search_result = heaps[0];
-    }
-
-    if (this->thread_pool_ != nullptr and search_thread_count > 1) {
-        for (auto& future : futures) {
-            future.get();
-        }
-        search_result = DistanceHeap::MakeInstanceBySize<true, true>(this->allocator_, topk);
-        for (auto& heap : heaps) {
-            auto size = heap->Size();
-            const auto* data = heap->GetData();
-            for (int i = 0; i < size; ++i) {
-                if (reasoning_ctx != nullptr and
-                    search_result->Size() >= static_cast<uint64_t>(topk) and
-                    data[i].first < search_result->Top().first) {
-                    reasoning_ctx->RecordEviction(search_result->Top().second / buckets_per_data_,
-                                                  1);
-                }
-                search_result->Push(data[i]);
-            }
-        }
-    }
-
-    // Deduplicate ids when buckets_per_data_ > 1
-    if (buckets_per_data_ > 1) {
-        std::unordered_map<InnerIdType, float> id_to_min_dist;
-        while (!search_result->Empty()) {
-            const auto& [dist_val, id] = search_result->Top();
-            auto origin_id = id / buckets_per_data_;
-            // Keep the smallest distance for each id
-            if (id_to_min_dist.find(origin_id) == id_to_min_dist.end() ||
-                dist_val < id_to_min_dist[origin_id]) {
-                id_to_min_dist[origin_id] = dist_val;
-            }
-            search_result->Pop();
-        }
-
-        auto cur_heap_top2 = std::numeric_limits<float>::max();
-        for (const auto& [origin_id, dist_val] : id_to_min_dist) {
-            if (dist_val < cur_heap_top2) {
-                search_result->Push(dist_val, origin_id);
-            }
-            if (search_result->Size() > origin_topk) {
-                search_result->Pop();
-            }
-            if (not search_result->Empty() and search_result->Size() == origin_topk) {
-                cur_heap_top2 = search_result->Top().first;
-            }
-        }
-    }
-
-    return search_result;
-}
-
-DistHeapPtr
-IVF::search_with_custom_distance(const DatasetPtr& query,
-                                 const SearchRequest& request,
-                                 const InnerSearchParam& param,
-                                 QueryContext& ctx,
-                                 ReasoningContext* reasoning_ctx) const {
-    const auto* query_data = query->GetFloat32Vectors();
-    Vector<BucketIdType> candidate_buckets(allocator_);
-    if (not param.bucket_ids.empty()) {
-        candidate_buckets.reserve(param.bucket_ids.size());
-        for (auto id : param.bucket_ids) {
-            candidate_buckets.push_back(static_cast<BucketIdType>(id));
-        }
-    } else {
-        candidate_buckets = partition_strategy_->ClassifyDatasForSearch(query_data, 1, param, &ctx);
-    }
-    if (reasoning_ctx != nullptr) {
-        reasoning_ctx->RecordBucketSelection(candidate_buckets);
-    }
-
-    int64_t topk = request.topk_;
-    const int64_t origin_topk = topk;
-    if (buckets_per_data_ > 1) {
-        CHECK_ARGUMENT(topk <= std::numeric_limits<int64_t>::max() / buckets_per_data_,
-                       "topk is too large for multi-bucket IVF search");
-        topk *= buckets_per_data_;
-    }
-
-    auto search_result = DistanceHeap::MakeInstanceBySize<true, false>(this->allocator_, topk);
-    const auto& filter = param.is_inner_id_allowed;
-    Filter* attr_filter = nullptr;
-
-    Vector<InnerIdType> candidate_ids(this->allocator_);
-    Vector<int64_t> candidate_labels(this->allocator_);
-    Vector<float> scores(this->allocator_);
-    const uint64_t batch_capacity = std::min<uint64_t>(
-        request.distance_batch_size_, std::max<uint64_t>(1, this->GetNumElements()));
-    candidate_ids.reserve(batch_capacity);
-    candidate_labels.reserve(batch_capacity);
-    scores.resize(batch_capacity);
-
-    auto is_timed_out = [&]() {
-        if (param.time_cost == nullptr or not param.time_cost->CheckOvertime()) {
-            return false;
-        }
-        if (ctx.stats != nullptr) {
-            ctx.stats->is_timeout.store(true, std::memory_order_relaxed);
-        }
-        return true;
-    };
-
-    auto submit_batch = [&]() {
-        if (candidate_ids.empty()) {
-            return true;
-        }
-        if (is_timed_out()) {
-            return false;
-        }
-        request.distance_batch_func_(
-            candidate_labels.data(), candidate_labels.size(), scores.data());
-        if (ctx.stats != nullptr) {
-            ctx.stats->AddDistance(SearchStatistics::DistancePhase::APPROXIMATE,
-                                   DistanceEvaluationBackend::UNKNOWN,
-                                   candidate_ids.size());
-        }
-        for (uint64_t i = 0; i < candidate_ids.size(); ++i) {
-            CHECK_ARGUMENT(std::isfinite(scores[i]),
-                           "custom query distance callback must return finite scores");
-            const auto origin_id = candidate_ids[i] / buckets_per_data_;
-            if (filter != nullptr and not filter->CheckValid(origin_id)) {
-                if (reasoning_ctx != nullptr) {
-                    reasoning_ctx->RecordFilterReject(origin_id);
-                }
-                continue;
-            }
-            if (reasoning_ctx != nullptr) {
-                reasoning_ctx->RecordVisit(origin_id, scores[i], 0);
-            }
-            search_result->Push(scores[i], candidate_ids[i]);
-            while (search_result->Size() > static_cast<uint64_t>(topk)) {
-                if (reasoning_ctx != nullptr) {
-                    reasoning_ctx->RecordEviction(search_result->Top().second / buckets_per_data_,
-                                                  0);
-                }
-                search_result->Pop();
-            }
-        }
-        candidate_ids.clear();
-        candidate_labels.clear();
-        return true;
-    };
-
-    bool timed_out = false;
-    for (const auto bucket_id : candidate_buckets) {
-        if (is_timed_out()) {
-            timed_out = true;
-            break;
-        }
-        if (bucket_id == INVALID_BUCKET_ID) {
-            continue;
-        }
-        if (not param.executors.empty()) {
-            param.executors[0]->Clear();
-            attr_filter = param.executors[0]->Run(bucket_id);
-        }
-        const auto bucket_size = bucket_->GetBucketSize(bucket_id);
-        const auto* ids = bucket_->GetInnerIds(bucket_id);
-        for (InnerIdType offset = 0; offset < bucket_size; ++offset) {
-            const auto inner_id = ids[offset];
-            if (inner_id == std::numeric_limits<InnerIdType>::max()) {
-                continue;
-            }
-            const auto origin_id = inner_id / buckets_per_data_;
-            if (attr_filter != nullptr and not attr_filter->CheckValid(offset)) {
-                if (reasoning_ctx != nullptr) {
-                    reasoning_ctx->RecordFilterReject(origin_id);
-                }
-                continue;
-            }
-            candidate_ids.push_back(inner_id);
-            candidate_labels.push_back(label_table_->GetLabelById(origin_id));
-            if (candidate_ids.size() == batch_capacity and not submit_batch()) {
-                timed_out = true;
-                break;
-            }
-        }
-        if (timed_out) {
-            break;
-        }
-    }
-    if (not timed_out) {
-        submit_batch();
-    }
-
-    if (buckets_per_data_ == 1) {
-        return search_result;
-    }
-
-    std::unordered_map<InnerIdType, float> id_to_min_score;
-    while (not search_result->Empty()) {
-        const auto& [score, inner_id] = search_result->Top();
-        const auto origin_id = inner_id / buckets_per_data_;
-        auto iter = id_to_min_score.find(origin_id);
-        if (iter == id_to_min_score.end() or score < iter->second) {
-            id_to_min_score[origin_id] = score;
-        }
-        search_result->Pop();
-    }
-
-    for (const auto& [origin_id, score] : id_to_min_score) {
-        search_result->Push(score, origin_id);
-        if (search_result->Size() > static_cast<uint64_t>(origin_topk)) {
-            search_result->Pop();
-        }
-    }
-    return search_result;
 }
 
 void
@@ -1872,369 +392,6 @@ IVF::check_merge_illegal(const vsag::MergeUnit& unit) const {
         throw VsagException(
             ErrorType::INVALID_ARGUMENT,
             "Merge Failed: IVF model not match, try to merge a different model ivf index");
-    }
-}
-
-DatasetPtr
-IVF::SearchWithRequest(const SearchRequest& request) const {
-    ValidateSearchThreshold(request.threshold_);
-    SearchStatistics stats;
-    QueryContext ctx{.alloc = request.search_allocator_, .stats = &stats};
-
-    bool is_range = (request.mode_ == SearchMode::RANGE_SEARCH);
-
-    auto param = this->create_search_param(request.params_str_, request.filter_);
-    const bool use_custom_distance = request.distance_batch_func_ != nullptr;
-    if (use_custom_distance) {
-        CHECK_ARGUMENT(request.distance_batch_size_ > 0,
-                       "custom query distance batch size must be greater than 0");
-        CHECK_ARGUMENT(not is_range, "IVF custom query distance only supports KNN search");
-        CHECK_ARGUMENT(not param.disable_bucket_scan,
-                       "IVF custom query distance does not support disable_bucket_scan");
-        CHECK_ARGUMENT(request.topk_ > 0, "topk must be greater than 0");
-        CHECK_ARGUMENT(param.parallel_search_thread_count == 1,
-                       "IVF custom query distance does not support parallel search");
-        param.enable_reorder = false;
-    }
-    param.query_context = &ctx;
-
-    if (not request.bucket_ids_.empty()) {
-        auto query_check = request.query_;
-        CHECK_ARGUMENT(query_check != nullptr, "query dataset cannot be null");
-        CHECK_ARGUMENT(request.bucket_ids_.size() == query_check->GetNumElements(),
-                       fmt::format("bucket_ids_ size must match the number of query vectors; "
-                                   "got {} bucket lists for {} queries",
-                                   request.bucket_ids_.size(),
-                                   query_check->GetNumElements()));
-        CHECK_ARGUMENT(query_check->GetFloat32Vectors() != nullptr,
-                       "query float32 vectors cannot be null");
-        CHECK_ARGUMENT(query_check->GetDim() == this->dim_,
-                       "query dimension must match index dimension");
-        CHECK_ARGUMENT(not param.disable_bucket_scan,
-                       "bucket_ids_ is incompatible with disable_bucket_scan mode");
-        for (uint64_t query_idx = 0; query_idx < request.bucket_ids_.size(); ++query_idx) {
-            const auto& ids = request.bucket_ids_[query_idx];
-            CHECK_ARGUMENT(not ids.empty(),
-                           fmt::format("bucket_ids_[{}] must not be empty; "
-                                       "use empty outer vector for default routing",
-                                       query_idx));
-            std::set<int64_t> seen_ids;
-            for (auto id : ids) {
-                CHECK_ARGUMENT(
-                    id >= 0,
-                    fmt::format(
-                        "bucket_id {} out of range [0, {})", id, this->bucket_->bucket_count_));
-                CHECK_ARGUMENT(
-                    id < static_cast<int64_t>(this->bucket_->bucket_count_),
-                    fmt::format(
-                        "bucket_id {} out of range [0, {})", id, this->bucket_->bucket_count_));
-                CHECK_ARGUMENT(seen_ids.insert(id).second,
-                               fmt::format("duplicate bucket_id {}", id));
-            }
-        }
-        if (query_check->GetNumElements() == 1) {
-            param.bucket_ids.assign(request.bucket_ids_[0].begin(), request.bucket_ids_[0].end());
-        }
-    }
-
-    auto query = request.query_;
-    if (use_custom_distance) {
-        CHECK_ARGUMENT(query != nullptr, "query dataset cannot be null");
-        CHECK_ARGUMENT(query->GetNumElements() == 1,
-                       "IVF custom search requires exactly one query");
-        CHECK_ARGUMENT(query->GetFloat32Vectors() != nullptr,
-                       "query float32 vectors cannot be null");
-        CHECK_ARGUMENT(query->GetDim() == this->dim_, "query dimension must match index dimension");
-    }
-    if (param.disable_bucket_scan) {
-        CHECK_ARGUMENT(query != nullptr, "query dataset cannot be null");
-        CHECK_ARGUMENT(query->GetNumElements() >= 1,
-                       "disable bucket scan requires at least one query");
-        CHECK_ARGUMENT(query->GetFloat32Vectors() != nullptr,
-                       "query float32 vectors cannot be null");
-        CHECK_ARGUMENT(query->GetDim() == this->dim_, "query dimension must match index dimension");
-        CHECK_ARGUMENT(not request.threshold_.has_value(),
-                       "threshold filtering is not supported with disable_bucket_scan");
-        auto result = this->route_buckets_only(query, param, ctx);
-        result->Statistics(stats.Dump());
-        return result;
-    }
-
-    if (query != nullptr && query->GetNumElements() > 1) {
-        CHECK_ARGUMENT(not is_range, "IVF batch search only supports KNN search");
-        CHECK_ARGUMENT(not use_custom_distance,
-                       "IVF batch search does not support custom query distance");
-        CHECK_ARGUMENT(request.expected_labels_.empty(),
-                       "IVF batch search does not support expected labels");
-        CHECK_ARGUMENT(request.topk_ > 0, "topk must be greater than 0");
-        CHECK_ARGUMENT(query->GetFloat32Vectors() != nullptr,
-                       "query float32 vectors cannot be null");
-        CHECK_ARGUMENT(query->GetDim() == this->dim_, "query dimension must match index dimension");
-
-        const auto num_queries = query->GetNumElements();
-        const auto total_slots = num_queries * request.topk_;
-        auto* alloc = select_query_allocator(ctx.alloc, this->allocator_);
-        auto* ids = static_cast<int64_t*>(alloc->Allocate(sizeof(int64_t) * total_slots));
-        auto* distances = static_cast<float*>(alloc->Allocate(sizeof(float) * total_slots));
-        std::fill_n(ids, total_slots, -1);
-        std::fill_n(distances, total_slots, std::numeric_limits<float>::infinity());
-
-        const auto* query_data = query->GetFloat32Vectors();
-        auto base_request = request;
-        base_request.expected_labels_.clear();
-        if (not request.bucket_ids_.empty()) {
-            base_request.bucket_ids_.clear();
-        }
-        auto search_func = [&](int64_t query_idx) -> void {
-            auto one_query = Dataset::Make();
-            one_query->NumElements(1)
-                ->Dim(query->GetDim())
-                ->Float32Vectors(query_data + query_idx * query->GetDim())
-                ->Owner(false);
-
-            auto one_request = base_request;
-            one_request.query_ = one_query;
-            if (not request.bucket_ids_.empty()) {
-                one_request.bucket_ids_ = {request.bucket_ids_[query_idx]};
-            }
-            JsonType json = JsonType::Parse(base_request.params_str_);
-            if (json.Contains("ivf")) {
-                json["ivf"]["parallelism"].SetInt64(1);
-            }
-            one_request.params_str_ = json.Dump();
-            auto one_result = this->SearchWithRequest(one_request);
-            const auto count = std::min(request.topk_, one_result->GetDim());
-            if (count > 0) {
-                std::copy_n(one_result->GetIds(), count, ids + query_idx * request.topk_);
-                std::copy_n(
-                    one_result->GetDistances(), count, distances + query_idx * request.topk_);
-            }
-        };
-
-        if (this->thread_pool_ != nullptr and param.parallel_search_thread_count > 1) {
-            std::vector<std::future<void>> futures;
-            for (int64_t query_idx = 0; query_idx < num_queries; ++query_idx) {
-                auto future = this->thread_pool_->GeneralEnqueue(search_func, query_idx);
-                futures.emplace_back(std::move(future));
-            }
-            for (auto& future : futures) {
-                future.get();
-            }
-        } else {
-            for (int64_t query_idx = 0; query_idx < num_queries; ++query_idx) {
-                search_func(query_idx);
-            }
-        }
-        auto result = Dataset::Make()
-                          ->NumElements(num_queries)
-                          ->Dim(request.topk_)
-                          ->Ids(ids)
-                          ->Distances(distances)
-                          ->Owner(true, alloc);
-        result->Statistics(stats.Dump());
-        return result;
-    }
-
-    if (request.enable_attribute_filter_ and this->attr_filter_index_ != nullptr) {
-        auto& schema = this->attr_filter_index_->field_type_map_;
-        auto expr = AstParse(request.attribute_filter_str_, &schema);
-        for (int64_t i = 0; i < param.parallel_search_thread_count; ++i) {
-            auto executor =
-                Executor::MakeInstance(this->allocator_, expr, this->attr_filter_index_);
-            executor->Init();
-            param.executors.emplace_back(executor);
-        }
-    }
-    std::shared_ptr<ReasoningContext> reasoning_ctx;
-    if (not request.expected_labels_.empty()) {
-        reasoning_ctx = std::make_shared<ReasoningContext>(this->allocator_);
-        reasoning_ctx->SetSearchParams(
-            request.topk_, "IVF", use_reorder_, request.filter_ != nullptr);
-
-        UnorderedMap<int64_t, InnerIdType> label_to_inner_id(this->allocator_);
-        std::vector<std::tuple<InnerIdType, BucketIdType, InnerIdType>> locations;
-        {
-            std::shared_lock<std::shared_mutex> lock(this->label_lookup_mutex_);
-            locations.reserve(request.expected_labels_.size());
-            for (const auto& label : request.expected_labels_) {
-                auto [success, inner_id] = this->label_table_->TryGetIdByLabel(label, true);
-                if (success) {
-                    label_to_inner_id[label] = inner_id;
-                    auto [bucket_id, offset_id] = this->get_location(inner_id);
-                    locations.emplace_back(inner_id, bucket_id, offset_id);
-                }
-            }
-        }
-
-        Vector<int64_t> expected_labels_vec(this->allocator_);
-        expected_labels_vec.reserve(request.expected_labels_.size());
-        for (const auto& label : request.expected_labels_) {
-            expected_labels_vec.push_back(label);
-        }
-        reasoning_ctx->InitializeExpectedTargets(expected_labels_vec, label_to_inner_id);
-
-        const auto* query_data = query->GetFloat32Vectors();
-        auto computer = this->bucket_->FactoryComputer(query_data);
-        for (const auto& [inner_id, bucket_id, offset_id] : locations) {
-            float dist = this->bucket_->QueryOneById(computer, bucket_id, offset_id);
-            if (ctx.stats != nullptr) {
-                ctx.stats->AddDistance(SearchStatistics::DistancePhase::APPROXIMATE,
-                                       this->bucket_->backend_);
-            }
-            reasoning_ctx->SetTrueDistance(inner_id, dist);
-        }
-        ctx.reasoning_ctx = reasoning_ctx.get();
-    }
-
-    if (use_custom_distance) {
-        param.search_mode = KNN_SEARCH;
-        param.topk = request.topk_;
-        auto search_result =
-            search_with_custom_distance(query, request, param, ctx, reasoning_ctx.get());
-        filter_search_result_by_threshold(
-            search_result, request.threshold_, select_query_allocator(ctx.alloc, this->allocator_));
-        if (search_result == nullptr || search_result->Empty()) {
-            auto dataset_results = DatasetImpl::MakeEmptyDataset();
-            this->AttachReasoningReport(dataset_results, reasoning_ctx.get());
-            dataset_results->Statistics(stats.Dump());
-            return dataset_results;
-        }
-        auto dataset_results = this->pack_knn_result(search_result, ctx.alloc);
-        this->AttachReasoningReport(dataset_results, reasoning_ctx.get());
-        dataset_results->Statistics(stats.Dump());
-        return dataset_results;
-    }
-
-    if (is_range) {
-        param.search_mode = RANGE_SEARCH;
-        param.radius = request.radius_;
-        param.range_search_limit_size = static_cast<int>(request.limited_size_);
-        if (use_reorder_ and param.enable_reorder and request.limited_size_ > 0) {
-            CHECK_ARGUMENT(param.factor > 0.0F,
-                           fmt::format("factor must be positive when use_reorder is true, got {}",
-                                       param.factor));
-            param.range_search_limit_size =
-                static_cast<int>(param.factor * static_cast<float>(request.limited_size_));
-        }
-        auto search_result = this->search<RANGE_SEARCH>(query, param, ctx, reasoning_ctx.get());
-        if (use_reorder_ and param.enable_reorder) {
-            int64_t k = (request.limited_size_ > 0) ? request.limited_size_
-                                                    : static_cast<int64_t>(search_result->Size());
-            auto result = reorder(
-                k, search_result, query->GetFloat32Vectors(), param, ctx, reasoning_ctx.get());
-            result->Statistics(stats.Dump());
-            this->AttachReasoningReport(result, reasoning_ctx.get());
-            return result;
-        }
-        auto dataset_results = this->pack_knn_result(search_result, ctx.alloc);
-        dataset_results->Statistics(stats.Dump());
-        this->AttachReasoningReport(dataset_results, reasoning_ctx.get());
-        return dataset_results;
-    }
-
-    // KNN mode
-    param.search_mode = KNN_SEARCH;
-    param.topk = request.topk_;
-    if (use_reorder_ and param.enable_reorder) {
-        CHECK_ARGUMENT(
-            param.factor > 0.0F,
-            fmt::format("factor must be positive when use_reorder is true, got {}", param.factor));
-        param.topk = static_cast<int64_t>(param.factor * static_cast<float>(request.topk_));
-        if (request.threshold_.has_value()) {
-            param.topk = std::max(param.topk, request.topk_);
-        }
-    }
-    const bool reorder_enabled = use_reorder_ and param.enable_reorder;
-    // Reordered searches defer the finite bound to exact distances, but bucket selection still
-    // needs threshold-mode state so non-finite approximations cannot consume the rerank pool.
-    param.distance_threshold = request.threshold_;
-    auto search_result = this->search<KNN_SEARCH>(query, param, ctx, reasoning_ctx.get());
-    if (reorder_enabled) {
-        auto result = reorder(request.threshold_.has_value() ? param.topk : request.topk_,
-                              search_result,
-                              query->GetFloat32Vectors(),
-                              param,
-                              ctx,
-                              reasoning_ctx.get(),
-                              request.threshold_);
-        result = FilterDatasetByThreshold(result, request.threshold_, ctx.alloc, request.topk_);
-        AttachReasoningReport(result, reasoning_ctx.get());
-        result->Statistics(stats.Dump());
-        return result;
-    }
-    filter_search_result_by_threshold(
-        search_result, request.threshold_, select_query_allocator(ctx.alloc, this->allocator_));
-    if (search_result == nullptr || search_result->Empty()) {
-        auto dataset_results = DatasetImpl::MakeEmptyDataset();
-        this->AttachReasoningReport(dataset_results, reasoning_ctx.get());
-        dataset_results->Statistics(stats.Dump());
-        return dataset_results;
-    }
-
-    auto dataset_results = this->pack_knn_result(search_result, ctx.alloc);
-    dataset_results->Statistics(stats.Dump());
-
-    this->AttachReasoningReport(dataset_results, reasoning_ctx.get());
-
-    return dataset_results;
-}
-
-void
-IVF::AttachReasoningReport(const DatasetPtr& dataset_results,
-                           ReasoningContext* reasoning_ctx) const {
-    if (reasoning_ctx == nullptr) {
-        return;
-    }
-    auto count = dataset_results->GetDim();
-    if (count > 0 and dataset_results->GetIds() != nullptr) {
-        Vector<InnerIdType> result_inner_ids(static_cast<uint64_t>(count), this->allocator_);
-        {
-            std::shared_lock<std::shared_mutex> lock(this->label_lookup_mutex_);
-            for (int64_t i = 0; i < count; ++i) {
-                result_inner_ids[i] =
-                    this->label_table_->GetIdByLabel(dataset_results->GetIds()[i]);
-            }
-        }
-        reasoning_ctx->MarkResult(result_inner_ids);
-    }
-    reasoning_ctx->DiagnoseExpectedTargets();
-    dataset_results->Reasoning(reasoning_ctx->GenerateReport());
-}
-
-void
-IVF::fill_location_map() {
-    this->location_map_.resize(this->total_elements_ * buckets_per_data_);
-    auto bucket_count = this->bucket_->bucket_count_;
-    if (precise_bucket_ != nullptr and precise_bucket_->GetBucketCount() != bucket_count) {
-        throw VsagException(ErrorType::INTERNAL_ERROR,
-                            "basic and precise bucket counts do not match");
-    }
-    for (BucketIdType i = 0; i < bucket_count; ++i) {
-        auto* ids = this->bucket_->GetInnerIds(i);
-        auto bucket_size = this->bucket_->GetBucketSize(i);
-        InnerIdType* precise_ids = nullptr;
-        if (precise_bucket_ != nullptr) {
-            auto precise_bucket_size = precise_bucket_->GetBucketSize(i);
-            if (precise_bucket_size != bucket_size) {
-                throw VsagException(ErrorType::INTERNAL_ERROR,
-                                    "basic and precise bucket sizes do not match");
-            }
-            precise_ids = precise_bucket_->GetInnerIds(i);
-        }
-        for (uint64_t j = 0; j < bucket_size; ++j) {
-            if (precise_ids != nullptr and precise_ids[j] != ids[j]) {
-                throw VsagException(ErrorType::INTERNAL_ERROR,
-                                    "basic and precise bucket inner ids do not match");
-            }
-            if (ids[j] == std::numeric_limits<InnerIdType>::max()) {
-                continue;
-            }
-            if (ids[j] >= this->total_elements_ * buckets_per_data_) {
-                throw VsagException(ErrorType::INTERNAL_ERROR, "invalid inner_id");
-            }
-            this->location_map_[ids[j] / buckets_per_data_] =
-                (static_cast<uint64_t>(i) << LOCATION_SPLIT_BIT) | static_cast<uint64_t>(j);
-        }
     }
 }
 
@@ -2487,31 +644,5 @@ IVF::GetMemoryUsage() const {
         memory += this->attr_filter_index_->GetMemoryUsage();
     }
     return memory;
-}
-
-void
-IVF::RebuildBucketGraphs() {
-    if (graph_build_threshold_ <= 0) {
-        throw VsagException(
-            ErrorType::UNSUPPORTED_INDEX_OPERATION,
-            "RebuildIVFBucketGraphs: index was not configured with graph_build_threshold > 0");
-    }
-    if (common_param_.data_type_ != DataTypes::DATA_TYPE_FLOAT) {
-        throw VsagException(ErrorType::UNSUPPORTED_INDEX_OPERATION,
-                            "RebuildIVFBucketGraphs: only supports float32 data type");
-    }
-
-    // Build new graphs in temp storage first (exception safety)
-    auto old_graphs = std::move(bucket_graphs_);
-    bucket_graphs_.resize(old_graphs.size(), nullptr);
-
-    try {
-        this->build_bucket_graphs();
-        this->cal_memory_usage();
-    } catch (...) {
-        // Restore old graphs on failure
-        bucket_graphs_ = std::move(old_graphs);
-        throw;
-    }
 }
 }  // namespace vsag
